adaptix/services/accounting/verify_accounting_placeholder.py
# This script triggers a Payroll Finalize API call and then checks the Accounting DB
import requests
import time
import psycopg2
import uuid
import json

HRMS_URL = "http://localhost:8000/api/hrms" # Via Kong
ACCOUNTING_DB = {
    "dbname": "accountingdb",
    "user": "postgres",
    "password": "passWORD", # We'll need to use container exec or adjust
    "host": "localhost",
    "port": "5448"
}

def run_test():
    print("🚀 Starting Accounting Automation Verification...")
    
    # 1. We assume Payroll Data exists from previous phase or we create it.
    # It's easier to verify via python script INSIDE hrms container to create data, then call API.
    # But better: Run `verify_payroll.py` again to generate a NEW finalized payslip? 
    # Or manually hit the API if we have IDs.
    
    print("⚠️  This test is best run manually or via integrated test suite.") 
    print("    Because we need valid Payroll IDs which change every run.")
    
    # Alternative: We inspect the Queue or Logs of accounting-worker.

if __name__ == "__main__":
    run_test()
